+++ src/main/java/com/hgys/iptv/model/bean/OrderProductDimensionDTO.java
@@ -5,6 +5,8 @@ import com.xuxueli.poi.excel.annotation.ExcelSheet;
 import lombok.Data;
 import org.apache.poi.hssf.util.HSSFColor;
 
+import java.math.BigDecimal;
+
 @ExcelSheet(name = "产品级单维度结算", headColor = HSSFColor.HSSFColorPredefined.LIGHT_GREEN)
 @Data
 public class OrderProductDimensionDTO {
@@ -38,6 +40,6 @@ public class OrderProductDimensionDTO {
      * 产品结算金额
      */
     @ExcelField(name = "结算金额（单位：元）")
-    private short money;
+    private BigDecimal money;
 
 }
